refactor: ScrollButton을 ScrollToTopButton으로 이름변경
- 폴더, 파일명 모두 변경
