+++ frontend/src/components/common/ScrollToTopButton/ScrollToTopButton.styles.ts
